#27 [REFACTOR] exception 패키지 domain 이동
