+++ src/main/java/com/example/project/application/user/UpdateUserService.java
@@ -1,6 +1,6 @@
 package com.example.project.application.user;
 
-import com.example.project.application.user.exception.UserNotFoundException;
+import com.example.project.domain.user.exception.UserNotFoundException;
 import com.example.project.application.user.usecase.UpdateUserCommand;
 import com.example.project.application.user.usecase.UpdateUserUseCase;
 import com.example.project.domain.user.model.User;
